webapp/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
attrs==21.4.0
    # via jsonschema
django==4.0.5
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular[sidecar]==0.22.1
    # via -r requirements.in
drf-spectacular-sidecar==2022.7.1
    # via drf-spectacular
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.6.1
    # via drf-spectacular
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
